+++ R/shorts_constructors.R
@@ -1,9 +1,10 @@
-new_shorts_model <- function(parameters, model_fit, model, data) {
+new_shorts_model <- function(parameters, model_fit, model, data, LOOCV = NULL) {
   shorts_model_object <- list(
     parameters = parameters,
     model_fit = model_fit,
     model = model,
-    data = data
+    data = data,
+    LOOCV = LOOCV
   )
 
   class(shorts_model_object) <- "shorts_model"
@@ -11,12 +12,13 @@ new_shorts_model <- function(parameters, model_fit, model, data) {
 }
 
 
-new_shorts_mixed_model <- function(parameters, model_fit, model, data) {
+new_shorts_mixed_model <- function(parameters, model_fit, model, data, LOOCV = NULL) {
   shorts_mixed_model_object <- list(
     parameters = parameters,
     model_fit = model_fit,
     model = model,
-    data = data
+    data = data,
+    LOOCV = LOOCV
   )
 
   class(shorts_mixed_model_object) <- "shorts_mixed_model"